+++ example/lib/main.dart
@@ -132,6 +132,9 @@ class _MyHomePageState extends State<MyHomePage>
         ),
         floatingActionButton: FloatingActionButton(
           onPressed: () {
+            if (sum is int && sum >= 10) {
+              sum = () => "above upperbound";
+            }
             switch (_tabController.index) {
               case 0:
                 int1++;