+++ example/lib/var.dart
@@ -21,7 +21,7 @@ int get int3 => _int3.value;
 set int3(value) => _int3.value = value;
 
 /// Computed Mediator Variable: sum
-final _sum = Computed(() => int1 + int2 + int3);
+final _sum = Rx(() => int1 + int2 + int3 as dynamic);
 get sum => _sum.value;
 set sum(value) => _sum.value = value;